+++ js/data.js
@@ -1,0 +1,57 @@
+const DESCRIPTION = [
+  'Летняя фотография на фоне лесной тропинки',
+  'ночная съемка',
+  'ночной город',
+  'спящий кот',
+  'полезный завтрак',
+  'хорошее завершение недели',
+  'трудовые будни',
+  'с друзьями на вечеринке',
+  'утро начинается с чашечки кофе',
+  'Падение Стены Мария',
+  'завтрак в Starbucks'
+];
+
+const NAMES = [
+  'Данил',
+  'Екатерина',
+  'Николай',
+  'Илья',
+  'Ярослав',
+  'Владислав',
+  'Валерия',
+  'Александр',
+  'Дмитрий',
+  'Виктория',
+  'Василиса',
+  'Ксения',
+  'Анна',
+  'Павел',
+  'Тимофей',
+  'Артем',
+  'Егор',
+  'Семён',
+  'Кира',
+  'Максим',
+  'Анна',
+  'Мария',
+  'Нина',
+  'Константин',
+  'Сергей',
+  'Виктор',
+  'Роман',
+  'Елизавета',
+  'Андрей',
+  'Леви Аккерман'
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export {DESCRIPTION, NAMES, COMMENTS};
